+++ libs/ardour/jack_slave.cc
@@ -34,7 +34,7 @@ using namespace sigc;
 JACK_Slave::JACK_Slave (jack_client_t* j)
 	: jack (j)
 {
-	float x;
+	double x;
 	nframes_t p;
 	/* call this to initialize things */
 	speed_and_position (x, p);
@@ -63,7 +63,7 @@ JACK_Slave::ok() const
 }
 
 bool 
-JACK_Slave::speed_and_position (float& sp, nframes_t& position) 
+JACK_Slave::speed_and_position (double& sp, nframes_t& position) 
 {
 	jack_position_t pos;
 	jack_transport_state_t state;
